+++ Source/FN.Application/Catalog/Product/Pattern/UpdateProductFacade.cs
@@ -3,7 +3,7 @@ using FN.Application.Systems.Redis;
 using FN.DataAccess;
 using FN.DataAccess.Entities;
 using FN.Utilities;
-using FN.ViewModel.Catalog.Products;
+using FN.ViewModel.Catalog.Products.Manage;
 using FN.ViewModel.Helper.API;
 using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
